requirements.txt
pygame
numpy
tensorflow
tf-agent
gym==0.10.11